Lots of modifications made to implement a User Sign-up

+++ src/main/java/com/autoworld/autoworld/controllers/CarController.java
@@ -3,16 +3,15 @@ package com.autoworld.autoworld.controllers;
 
 import com.autoworld.autoworld.models.Car;
 import com.autoworld.autoworld.models.Customer;
+import com.autoworld.autoworld.models.User;
 import com.autoworld.autoworld.models.data.CarDao;
 import com.autoworld.autoworld.models.data.CustomerDao;
+import com.autoworld.autoworld.models.data.UserDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -26,11 +25,16 @@ public class CarController {
     @Autowired
     private CustomerDao customerDao;
 
+    @Autowired
+    private UserDao userDao;
+
 
 
 
     @RequestMapping(value = "")
-    public String displayCars(Model model) {
+    public String displayCars(Model model, @CookieValue(value = "user", defaultValue = "none") String username) {
+
+        if(username.equals("none")) { return "redirect:/user/login"; }
         model.addAttribute("cars", carDao.findAll());
         model.addAttribute("title", "Current Cars");
 
@@ -38,7 +42,8 @@ public class CarController {
     }
 
     @RequestMapping(value = "add", method = RequestMethod.GET)
-    public String addCarForm(Model model) {
+    public String addCarForm(Model model, @CookieValue(value = "user", defaultValue = "none") String username) {
+        if(username.equals("none")) { return "redirect:/user/login"; }
         model.addAttribute("title", "Add a New Car");
         model.addAttribute(new Car());
         model.addAttribute("customers", customerDao.findAll());
@@ -47,7 +52,10 @@ public class CarController {
 
     @RequestMapping(value = "add", method = RequestMethod.POST)
     public String processAddCarForm(@ModelAttribute @Valid Car newCar,
-                                    Errors errors, Model model, @RequestParam int customerId) {
+                                    Errors errors, Model model, @RequestParam int customerId,
+                                    @CookieValue(value = "user", defaultValue = "none") String username) {
+        if(username.equals("none")) { return "redirect:/user/login"; }
+
         if (errors.hasErrors()) {
             model.addAttribute("title", "Add a New Car");
             model.addAttribute("title", customerDao.findAll());
@@ -60,13 +68,15 @@ public class CarController {
     }
 
     @RequestMapping(value = "delete", method = RequestMethod.GET)
-    public String displayDeleteCarForm(Model model) {
+    public String displayDeleteCarForm(Model model, @CookieValue(value = "user", defaultValue = "none") String username) {
+        if(username.equals("none")) { return "redirect:/user/login"; }
         model.addAttribute("cars", carDao.findAll());
         model.addAttribute("title", "Delete Car(s)");
         return "cars/delete";
     }
     @RequestMapping(value = "delete", method = RequestMethod.POST)
-    public String processDeleteCarForm(@RequestParam int[] carIds) {
+    public String processDeleteCarForm(@RequestParam int[] carIds, @CookieValue(value = "user", defaultValue = "none") String username) {
+        if(username.equals("none")) { return "redirect:/user/login"; }
 
         for (int carId : carIds) {
             carDao.delete(carId);
